+++ part2/countries/src/App.js
@@ -1,81 +1,9 @@
 
 import React, { useState, useEffect } from 'react'
 import axios from 'axios'
+import Filter from './components/Filter'
+import List from './components/List'
 
-const Filter = ({ handleFilterCountry }) => (
-  <>
-    <p> Find countries </p>
-    <input onChange={handleFilterCountry} />
-  </>
-)
-
-const Language = ({ language }) => (
-  <li>{language}</li>
-)
-
-
-const Country = ({ country }) => {
-  const [weather, setWeather] = useState([])
-  // const api_key = process.env.REACT_APP_API_KEY
-
-
-  useEffect(() => {
-    axios
-      .get(`http://api.weatherapi.com/v1/current.json?key=${process.env.REACT_APP_API_KEY}&q=${country.capital}`)
-      .then(response => {
-        setWeather(response.data.current.condition)
-      })
-  }, [country])
-
-  return (
-    <>
-      <h3>{country.name}</h3>
-      <p>Capital: {country.capital}</p>
-      <p>Population: {country.population}</p>
-      <h3>Language(s)</h3>
-
-      <ul>
-        {country.languages.map((language) =>
-          <Language key={language.name} language={language.name} />
-        )}
-      </ul>
-
-      <img alt="flag" src={country.flag} />
-
-      <h3>Weather in {country.capital}</h3>
-      <div>{weather.text}</div>
-      <img alt="flag" src={weather.icon} />
-    </>
-  )
-}
-
-const List = ({ countriesList, setToCountry }) => {
-  // joprojām šis if neiet cauri
-  if (countriesList.length > 10) {
-    return (
-      <p>Too many matches, specify another filter</p>
-    )
-  } else if (countriesList.length === 1) {
-    return (
-      <>
-        {countriesList.map((country) =>
-          <Country key={country.name} country={country} />
-        )}
-      </>
-    )
-  } else if (countriesList.length <= 10) {
-    return (
-      <ul>
-        {countriesList.map(country =>
-          <li key={country.name}>{country.name} <button onClick={() => setToCountry(country.name)}>Show</button></li>
-        )}
-      </ul>
-    )
-  }
-  return (
-    <p>Feel free to search for a country</p>
-  )
-}
