+++ part2/countries/src/components/Filter.js
@@ -1,0 +1,10 @@
+import React from 'react'
+
+const Filter = ({ handleFilterCountry }) => (
+    <>
+        <p> Find countries </p>
+        <input onChange={handleFilterCountry} />
+    </>
+)
+
+export default Filter
