+++ part2/countries/src/components/Language.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const Language = ({ language }) => (
+    <li>{language}</li>
+)
+
+export default Language
